RDM-2811 - moved classes to security package
