--- src/main/java/uk/gov/hmcts/ccd/domain/service/search/elasticsearch/CaseSearchRequestSecurity.java
@@ -1,7 +1,0 @@
-package uk.gov.hmcts.ccd.domain.service.search.elasticsearch;
-
-public interface CaseSearchRequestSecurity {
-
-    void secureRequest(CaseSearchRequest caseSearchRequest);
-
-}
