+++ src/main/java/uk/gov/hmcts/ccd/domain/service/search/elasticsearch/security/CaseSearchRequestSecurity.java
@@ -1,0 +1,9 @@
+package uk.gov.hmcts.ccd.domain.service.search.elasticsearch.security;
+
+import uk.gov.hmcts.ccd.domain.service.search.elasticsearch.CaseSearchRequest;
+
+public interface CaseSearchRequestSecurity {
+
+    void secureRequest(CaseSearchRequest caseSearchRequest);
+
+}
