+++ src/main/java/uk/gov/hmcts/ccd/domain/service/search/elasticsearch/security/ElasticsearchUserCaseAccessFilter.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.ccd.domain.service.search.elasticsearch;
+package uk.gov.hmcts.ccd.domain.service.search.elasticsearch.security;
 
 import java.util.List;
 import java.util.Optional;
