+++ src/test/java/uk/gov/hmcts/ccd/domain/service/search/elasticsearch/ElasticsearchUserCaseAccessFilterFactoryTest.java
@@ -18,6 +18,7 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import uk.gov.hmcts.ccd.domain.service.common.CaseAccessService;
+import uk.gov.hmcts.ccd.domain.service.search.elasticsearch.security.ElasticsearchUserCaseAccessFilter;
 
 class ElasticsearchUserCaseAccessFilterFactoryTest {
